Peggy

Packrat parser combinators for Julia supporting left-recursive grammars.

Features:

Pretty good error messages

pretty good syntax error messages. [TODO: how costly is this? Make it optional?]

detects and correctly handles left-recursion

separation of syntax from functinality. I'm still playing around with the input syntax.

Quickstart

Creating Parsers

Use @grammar to construct parsers.

julia> using Peggy

julia> p = @grammar begin
          start = "match a"
       end;

julia> runpeg(p, "match abc") # does not need to match the entire string
"match a"

julia> p = @grammar begin
          start = [ a_or_bs " "... hexdigit !anych() ]
          a_or_bs = ("a" / "b")...
          hexdigit = anych("[:digit:]a-fA-F")
       end;

julia> p("abaab 9")
(a_or_bs = ["a", "b", "a", "a", "b"], hexdigit = "9")

Index

Peggy.END

Peggy.NonemptyRegex

Peggy.Parser

Peggy.anych

Peggy.anych

Peggy.grammar

Peggy.many

Peggy.not

Peggy.oneof

Peggy.pegparser

Peggy.runpeg

Peggy.@grammar

Peggy.END — Constant

A PEG parser that matches the end of the input; yields result ().

Peggy.NonemptyRegex — Type

Succeeds with a non-empty value.

Peggy.Parser — Method

(parser::Parser)(input)

Parsers are callable.

Peggy.anych — Method

anych(charclass::String)

Create a parser for a single character matchng regex character classes.

Functionally identical to Regex("[charclass]") except it is known to never match an empty string. This is important to avoid unneccesary and expensive left-recursion overhead.

Examples

julia> g = @grammar begin
       number = [ digit ds:(digit...)  { parse(Int, *(digit, ds...)) } ]
       digit = anych("[:digit:]")
       end;

julia> g("1234")
1234

Peggy.anych — Method

A parser that matches any single character.

Peggy.grammar — Method

grammar([start::Symbol], (symbol => peg_expr)...)

Create a parser from a set of productions, which are named, mutually recursive parsers.

Parsers that are members of a grammar can reference are member parser by their symbol.

If start is omitted, the symbol of the first production is used.

Peggy.many — Method

many(pegexpr...)

Create a parser that matches zero or more repititions of parser(pegexpr); return a vector of results.

Peggy.not — Method

not(p)

Create a parser that fails if parser p succeeds. Otherwise it succeeds with value ()

Peggy.oneof — Method

oneof(pegexpr...)

Create a parser for ordered alternatives.

Peggy.pegparser — Method

pegparser(peg_expr)

Create a Parser from a PEG expression.

PEG Expressions

- `String` - matches & yields the string literal
- `Regex` - matches the  `Regex` and yields the match value
- `Symbol` - matches to expression associated with the symbol in a `grammar`.
- `[peg_expr...]` - matches the each expression in sequence. Yield a tuple of the result
- `peg_exr...` - short-hand for the above, where syntatically valid.
- `peg_expr => callable` - matches `peg_expr` and yields result of applying `callable` to its value.
- `peg_epxr => k` - short-hand for `peg_expr` => _ -> k`
- `Parser` - the result of `onoeof`, etc.

See also

Any PEG parser is also a valid PEG expression There are a few other combinators that generate PEG parsers.

- [`oneof(peg_expr...)`](@ref) - yields value of the first matching `peg_epxr`
- [`many(peg_expr)`](@ref) - matches zero or more, yields a `AbstractVector` of the results
- [`not(peg_expr)`](@ref) - matches only if `peg_epxr` fails. Yields `()`

Peggy.runpeg — Method

runpeg(e::Parser, s::AbstractString) => result_value

Parse the input with the parser.

Returns the resulting value or throws a ParseException.

Peggy.@grammar — Macro

Create a PEG parser from grammar expression.

@grammar begin
    grammar_expression = peg_expression
    peg_expression = [
        literal / regex_character_class
    ]
    regex_character_class = anych("[:alpha:]_") 
    literal = "an string"
end;