Peggy

Generate Packrat PEG parsers for Julia

Features:

pretty good syntax error messages.

detects indirect left-recursive rules during compilation. Only left-recursive rules pay a performance cost

both combinator functions and a macro are provided

A Peggy.Parser is function that takes a string as input and returns its parsed value.

Create parsers using either a succinct Peggy expression via the @peg macro or lower-level functions.

Index

Peggy.Parser

Peggy.fail

Base.:!

Base.:*

Base.:|

Peggy.ANY

Peggy.CHAR

Peggy.END

Peggy.followedby

Peggy.grammar

Peggy.many

Peggy.not

Peggy.oneof

Peggy.pegparser

Peggy.runpeg

Peggy.@peg

Peggy.Parser — Type

(parser::Parser)(input)

A Peggy.Parser is function that takes a string as input and returns its parsed value.

Peggy.fail — Type

fail(message) => Parser

A parser that always fails with the given message.

Useful for error messages.

Base.:! — Method

!(p::Parser) == not(p)

Base.:* — Method

p::Parser * n == many(p; min=n)
p::Parser * (a:b) == many(p; min=a, max=b)

Base.:| — Method

 p1::Parser | p2::Parser == oneof(p1, p2)

A short-form for oneof.

Peggy.ANY — Method

A PEG parser that matches any character and yields it as a string.

Peggy.CHAR — Method

CHAR(charclass::String)

Create a parser for a single character matchng regex character classes.

Functionally identical to Regex("[charclass]") except it is known to never match an empty string. This is important to avoid unneccesary and expensive left-recursion overhead.

Examples

julia> g = @grammar begin
       number = [ digit ds:(digit...)  { parse(Int, *(digit, ds...)) } ]
       digit = CHAR("[:digit:]")
       end;

julia> g("1234")
1234

Peggy.END — Method

A PEG parser that matches the end of the input; yields result ().

Peggy.followedby — Method

followedby(expr...)
!!(e::Parser)

Create a parser that matches expr but consumes nothing.

Peggy.grammar — Method

grammar([start::Symbol], (symbol => peg_expr)...)

Create a parser from a set of productions, which are named, mutually recursive parsers.

Parsers that are members of a grammar can reference are member parser by their symbol.

If start is omitted, the symbol of the first production is used.

Peggy.many — Method

many(exprs...; min=0, max=missing)

Create a parser that matches zero or more repititions of the sequence expr...; returns a vector of results.

Peggy.not — Method

not(p)

Create a parser that fails if parser p succeeds. Otherwise it succeeds with value ()

Peggy.oneof — Method

oneof(pegexpr...)

Create a parser for ordered alternatives.

Peggy.pegparser — Method

pegparser(peg_expr)

Create a Parser from a PEG expression.

PEG Expressions

- `String` - matches & yields the string literal
- `Regex` - matches the  `Regex` and yields the match value
- `Symbol` - matches to expression associated with the symbol in a `grammar`.
- `[peg_expr...]` - matches the each expression in sequence. Yield a tuple of the result
- `peg_exr...` - short-hand for the above, where syntatically valid.
- `peg_expr => callable` - matches `peg_expr` and yields result of applying `callable` to its value.
- `peg_epxr => k` - short-hand for `peg_expr` => _ -> k`
- `Parser` - the result of `onoeof`, etc.

See also

Any PEG parser is also a valid PEG expression There are a few other combinators that generate PEG parsers.

- [`oneof(peg_expr...)`](@ref) - yields value of the first matching `peg_epxr`
- [`many(peg_expr)`](@ref) - matches zero or more, yields a `AbstractVector` of the results
- [`not(peg_expr)`](@ref) - matches only if `peg_epxr` fails. Yields `()`

Peggy.runpeg — Method

runpeg(e::Parser, s::AbstractString) => result_value

Parse the input with the parser.

Returns the resulting value or throws a ParseException.

Peggy.@peg — Macro

Create a Peggy.Parser from a Peggy expression